Test_2/requirements.txt
torch
torchvision
transformers
datasets
Pillow
matplotlib
seaborn